+++ test/src/Queue-test.js
@@ -1146,7 +1146,7 @@ describe('Queue', () => {
 
       queue.queueMessage({routingKey: 'test.3'});
 
-      expect(queue.get()).to.be.false;
+      expect(queue.get()).to.be.undefined;
       const consumer = queue.consume(() => {});
 
       expect(queue.messageCount).to.equal(2);