+++ types/Exchange.d.ts
@@ -29,7 +29,7 @@ export interface Exchange {
   getState(): any;
   stop(): void;
   recover(state: any, getQueue: CallableFunction): Exchange;
-  getBinding(queueName: string, pattern: string): Binding | undefined;
+  getBinding(queueName: string, pattern: string): Binding;
   emit(eventName: string, content?: any): any;
   on(pattern: string, handler: CallableFunction, consumeOptions?: {consumerTag?: string, [x: string]: any}): Consumer;
   off(pattern: string, handler: any): void;